fix: dependency resolution issue resolved in skills
