+++ docs/api/skills/abstract_round_abci/test_tools/integration.md
@@ -79,12 +79,7 @@ the incoming message
 #### process`_`n`_`messages
 
 ```python
-def process_n_messages(ncycles: int, synchronized_data: Union[
-            TxSettlementSynchronizedSata,
-            PriceEstimationSynchronizedSata,
-            LiquidityRebalancingSynchronizedSata,
-            None,
-        ] = None, behaviour_id: Optional[str] = None, handlers: Optional[HandlersType] = None, expected_content: Optional[ExpectedContentType] = None, expected_types: Optional[ExpectedTypesType] = None, fail_send_a2a: bool = False, mining_interval_secs: float = 0) -> Tuple[Optional[Message], ...]
+def process_n_messages(ncycles: int, synchronized_data: Optional[BaseSynchronizedData] = None, behaviour_id: Optional[str] = None, handlers: Optional[HandlersType] = None, expected_content: Optional[ExpectedContentType] = None, expected_types: Optional[ExpectedTypesType] = None, fail_send_a2a: bool = False, mining_interval_secs: float = 0) -> Tuple[Optional[Message], ...]
 ```
 
 Process n message cycles.
@@ -105,88 +100,6 @@ Process n message cycles.
 
 tuple of incoming messages
 
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._SafeConfiguredHelperIntegration"></a>
-
-## `_`SafeConfiguredHelperIntegration Objects
-
-```python
-class _SafeConfiguredHelperIntegration(IntegrationBaseCase)
-```
-
-Base test class for integration tests with Gnosis, but no contract, deployed.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._SafeConfiguredHelperIntegration.setup"></a>
-
-#### setup
-
-```python
-@classmethod
-def setup(cls, **kwargs: Any) -> None
-```
-
-Setup.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._GnosisHelperIntegration"></a>
-
-## `_`GnosisHelperIntegration Objects
-
-```python
-class _GnosisHelperIntegration(_SafeConfiguredHelperIntegration)
-```
-
-Class that assists Gnosis instantiation.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._GnosisHelperIntegration.setup"></a>
-
-#### setup
-
-```python
-@classmethod
-def setup(cls, **kwargs: Any) -> None
-```
-
-Setup.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._TxHelperIntegration"></a>
-
-## `_`TxHelperIntegration Objects
-
-```python
-class _TxHelperIntegration(_GnosisHelperIntegration)
-```
-
-Class that assists tx settlement related operations.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._TxHelperIntegration.sign_tx"></a>
-
-#### sign`_`tx
-
-```python
-def sign_tx() -> None
-```
-
-Sign a transaction
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._TxHelperIntegration.send_tx"></a>
-
-#### send`_`tx
-
-```python
-def send_tx(simulate_timeout: bool = False) -> None
-```
-
-Send a transaction
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration._TxHelperIntegration.validate_tx"></a>
-
-#### validate`_`tx
-
-```python
-def validate_tx(simulate_timeout: bool = False, mining_interval_secs: float = 0) -> None
-```
-
-Validate the sent transaction.
-
 <a id="packages.valory.skills.abstract_round_abci.test_tools.integration._HarHatHelperIntegration"></a>
 
 ## `_`HarHatHelperIntegration Objects
@@ -198,50 +111,6 @@ class _HarHatHelperIntegration(IntegrationBaseCase)
 Base test class for integration tests with HardHat provider.
 
 <a id="packages.valory.skills.abstract_round_abci.test_tools.integration._HarHatHelperIntegration.setup"></a>
-
-#### setup
-
-```python
-@classmethod
-def setup(cls, **kwargs: Any) -> None
-```
-
-Setup.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration.GnosisIntegrationBaseCase"></a>
-
-## GnosisIntegrationBaseCase Objects
-
-```python
-class GnosisIntegrationBaseCase(
-    _TxHelperIntegration,  _HarHatHelperIntegration,  HardHatAMMBaseTest)
-```
-
-Base test class for integration tests in a Hardhat environment, with Gnosis deployed.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration.GnosisIntegrationBaseCase.setup"></a>
-
-#### setup
-
-```python
-@classmethod
-def setup(cls, **kwargs: Any) -> None
-```
-
-Setup.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration.AMMIntegrationBaseCase"></a>
-
-## AMMIntegrationBaseCase Objects
-
-```python
-class AMMIntegrationBaseCase(
-    _TxHelperIntegration,  _HarHatHelperIntegration,  HardHatAMMBaseTest)
-```
-
-Base test class for integration tests in a Hardhat environment, with AMM interaction.
-
-<a id="packages.valory.skills.abstract_round_abci.test_tools.integration.AMMIntegrationBaseCase.setup"></a>
 
 #### setup
