+++ packages/hashes.csv
@@ -4,16 +4,16 @@ open_aea/protocols/scaffold,bafybeihhyjuqvp5zqtikkjrwtikg4fr4tsub43n777huxtpeu77
 open_aea/protocols/signing,bafybeid6f5ceool4evgaxdegs5pkjoivqjk3mo3ehctalswngb5m5c32wm
 open_aea/skills/scaffold,bafybeih3h7yia76rrbittnoeotpraiuho5rdqsnksuvncinaa3akgmnbaa
 valory/agents/abstract_abci,bafybeifc3xhwg6yhc25al3coc5ugngiqc5zygfpicl7ntpwsrzxt3gecd4
-valory/agents/apy_estimation,bafybeiacs2kuowwshc7zftf6u5uy5g5iftqp72br75qgd3ey7hczjvftm4
-valory/agents/apy_estimation_chained,bafybeia7lillchql7osgksvz7jbzs5a27twjpqe3i4p6mun4qhpcjp5pbq
+valory/agents/apy_estimation,bafybeiemzfzmgo5sk222miw4zwc2gnt7wdqsub6ddguwrfukq3pwiaubci
+valory/agents/apy_estimation_chained,bafybeicus2nrolpkn67qft2deymwrzklze62esreq36tnzuyf3fqae7roa
 valory/agents/counter,bafybeibvok2be6avmhrafwahcdjaq4hszhjym6dytsctsqcxmvjiko67lu
 valory/agents/counter_client,bafybeigkvd5i2jvpj4z5dfyzccqttfuxdwztabngi2bjkxdkdv2rb55kua
-valory/agents/hello_world,bafybeigsg6aikfzr5dzhdivxxsqxuxsazihjtmmnsr6y6dpdt5mxgdf6iy
-valory/agents/oracle,bafybeibjnqewgxhxzbqafvqnxmq2ilwhfv6brhz5wtjypmjg74heuyfpgq
-valory/agents/register_reset,bafybeiedssyqltgi3wvvrvac7s3zmwtnmpezk5qlxtraw2konafzf6bktq
-valory/agents/registration_start_up,bafybeih3dzm2hyrsfbprdm4aemfk2dprlhowbhuihcs236b6wruci6gtou
-valory/agents/simple_abci,bafybeif5zwfvpr57s64xee5jef62o3kjpbwe65ltfir5xwpz77bpx7qbhm
-valory/agents/test_abci,bafybeicx5bl4ledjkjx5kxxxi6kqcyescy3a2cy24jipdtk2rrwjsq2sxm
+valory/agents/hello_world,bafybeiceeubzebfnj6x6mdac6kpn2rtdk3xea6lix4iboiicx5qpliapg4
+valory/agents/oracle,bafybeiggp7g3reb3q3hisvgjqwgvforbhz26caraqi5wi4cqzny5uiezcy
+valory/agents/register_reset,bafybeifkvg2paaso5bqtwsrs4623g55j6wtslop74rjhyvrw2vteatbl7e
+valory/agents/registration_start_up,bafybeiaaxhr6wavs35d2twdvvp6sad5yj4qu5ho3wke7qjzxz44ckvyz4e
+valory/agents/simple_abci,bafybeiey3nwybt3nkrb4chiuto72t4rxhqfciukfchzruu7d4etnk5txa4
+valory/agents/test_abci,bafybeidoi6nbfln7zticioxm6uh4rr35ngrlxgiiklgj5vbgmhxbnlz47m
 valory/connections/abci,bafybeibfdzfevrtiywolqeepk6yfzyvl4bbu6mghhnebkptcbhsh5fxkbu
 valory/connections/http_client,bafybeicmoolrrjnt5qn7jbwon7a3b5noe6hkj7kbjj57xtfosoygkckhuy
 valory/connections/ledger,bafybeibskbi7lft2htvogxrwzbix6kj2ri373ppde6bkmuabmq3z32lkxi
@@ -31,31 +31,31 @@ valory/protocols/contract_api,bafybeigkedpmqhm7lrqusnkyfcapyx3dlkz2djfq74hou5pwt
 valory/protocols/http,bafybeienmids6f5pcey2bnywmiup2faawqgmdhhz65qfciuhkntugztvpi
 valory/protocols/ledger_api,bafybeif2c3jt4tim3zbfq5c2lds7bs6qj5gvj2obk3pzdk3uy6zk2ivo3e
 valory/protocols/tendermint,bafybeihcnjhovvyyfbkuw5sjyfx2lfd4soeocfqzxz54g67333m6nk5gxq
-valory/services/apy_estimation_hardhat,bafybeiawce4gk2rmdvus26gzfmirboadkmci5swknduh2wbdsb2kt26yy4
+valory/services/apy_estimation_hardhat,bafybeigg5fvjatq5zs3jp4zgfbv5gm7k7o4bh42jdjkuof6ug2d5koar4a
 valory/services/counter,bafybeidahelchmcwkymllanimjlhinrtrfii2lzhbwiw4h2rj6tti72ygi
-valory/services/hello_world,bafybeibky2hr7ykopkqpedccttlhca7nhf7nyj6vmihhqcn7cyv2jlggpi
-valory/services/oracle_goerli,bafybeifars5trm3ajygy4eywhvz2irf4eg2yazb3q2kpsly2apnond555m
-valory/services/oracle_hardhat,bafybeigowobnbd4hfcov3yzyqutxzocsntwjypyu6bib22h2fmiaznacpi
-valory/services/oracle_polygon,bafybeiaxqv7ak3v2w46yarectn7dlenned56genhodlisisnzg476ikv44
-valory/services/oracle_ropsten,bafybeiawneiozrm7oly5cl2kabw4noi4x64dfmgb6f57e5fk4bzdohzekq
-valory/services/oracle_ropsten_cluster,bafybeielcr2u3htdwdfhgmcl3ttpp27iszu4udnnhs325abrshcmz7247q
-valory/services/register_reset,bafybeiamz7vpvo5dwg7s444wvbeinhfb3d37wg42on54emkpepi3f75dqq
+valory/services/hello_world,bafybeihatxt32tqhkdemrnzwz326xxul7jisbqo2crclibavmxjzm2v7wu
+valory/services/oracle_goerli,bafybeihjlqn3saweb6he3pwnjf5yijhcaudwgglc5x7dw2ykukzqlfexea
+valory/services/oracle_hardhat,bafybeiectmvonkmrmder2hnmeji2e26q5hszwjnqxgucu5umvsv3tcvhee
+valory/services/oracle_polygon,bafybeiahtyxgnkmo64p2tqkpfaef6qih5n5moxcb7uzmhoivirfotb5d6y
+valory/services/oracle_ropsten,bafybeihcfvff2q7mfsb5jcfstwmo5xivvvcftit35amr34ikrp5ompr4yy
+valory/services/oracle_ropsten_cluster,bafybeib5ubr66qzvajwpaueye3up3isla7md4l2tyiczobnynsc5kscf3a
+valory/services/register_reset,bafybeihp5e65rwyyqlbod4t75w3dxczvcjuj6fg3q2qlahstvldhaoqal4
 valory/skills/abstract_abci,bafybeih62tn62c33bcffjrlxvvokvorli2ow4v6myij352bl3im5dhvloa
-valory/skills/abstract_round_abci,bafybeievrok3xa3foeay3d3xoscunhbgnqiu6jyse6yck4td76dxsbvsiu
-valory/skills/apy_estimation_abci,bafybeibm75x73ywdfo7o7lykrjtn5ip6woi3qmevrq63k33lfdrcd37kdi
-valory/skills/apy_estimation_chained_abci,bafybeih7rchwopfe5einr6ahbldtq64acwuvwpe353hr2j6i7v4nz5rkea
+valory/skills/abstract_round_abci,bafybeihcarapqbgd7czlfyygkdjvb3zxhfkxvud7hd5owajkcwqsxa63gy
+valory/skills/apy_estimation_abci,bafybeiatalnnmdu5o7ewi3jttryyg2wnbrfjxczwnaxwtslvdsfe7lridy
+valory/skills/apy_estimation_chained_abci,bafybeiejkky4sopg7flnxs5pvf5ksprtij3iuuib53a3oviq5shis4pxcy
 valory/skills/counter,bafybeiho75vypyciiyrmi52zkhlaz2jjgn7vaxpoztncnm76ljtltuxea4
 valory/skills/counter_client,bafybeidyuyencl4mg6ph3ufxisxgkfh2lavecik3lq5sbdzjyctg5jv4yy
-valory/skills/hello_world_abci,bafybeifp2vpwiapnusrrpi6mjedxgmovodyrdc7mgvqezhq3xcaapaqfom
-valory/skills/liquidity_provision_abci,bafybeihy7xsh5sxvcsqsejjkalouir3ltyjik72gasngixbay6jbdwpl6q
-valory/skills/liquidity_rebalancing_abci,bafybeidbdq6ofcwpfcmy2npcybteoohs6mpxcrvlgmkolsm5onkkh4rr6i
-valory/skills/oracle_abci,bafybeiclvchyb5jstccepgg7ltiiq3xuh3q5fqd4z5rmi7ugrpvwt5tgg4
-valory/skills/oracle_deployment_abci,bafybeicobpm5puiw5wcjbj5yxb4qwnm2ceu65bxepzhphwr7mnz7oorv5m
-valory/skills/price_estimation_abci,bafybeifhypfl22b4pwusduxnpd7i4stamz7w36yxluycjbpo5gm7tnuj5q
-valory/skills/register_reset_abci,bafybeifsojtegc2ecbrvn4hvn4zotpdxn75lasklr5c56o67afqtvulycy
-valory/skills/registration_abci,bafybeia3g5vjngmccglc7yx2mx6l4hh76fhe7ot2plqrd36lllekrsznae
-valory/skills/reset_pause_abci,bafybeifcgdaugnmv7qcmhdxhcfbemnsatwzmn6yxcl2b32qovyluojnu2u
-valory/skills/safe_deployment_abci,bafybeiahpjlxxcyelayvmpgd4fqz4j7hja7wdnwjmdnij32455izjmkawy
-valory/skills/simple_abci,bafybeicm6iezcbysidoscjtevcwozqlfcxwrdg42ldlc4s2owcgpq7d73e
-valory/skills/test_abci,bafybeihypwoinm2vhdees6xrrdpmrsd3lewkkx32oryvetdyo75dnlkarm
-valory/skills/transaction_settlement_abci,bafybeiavtszkfhm7fly7ro447rfioitzlj5gtniubumlvz32zw2ahynnpa
+valory/skills/hello_world_abci,bafybeiflibb7eiydyxuoyvsasc4sq5n3cpyplboprkngu5yhqsncfaicky
+valory/skills/liquidity_provision_abci,bafybeibfvti7fqjk2bxe4ccw4kk3jfdduepvrfn25d7wmpmww6xl4vombi
+valory/skills/liquidity_rebalancing_abci,bafybeibaanttagxccjzohupn5syfezfov777uwbiwlioz2jqc3c6t5jmz4
+valory/skills/oracle_abci,bafybeiaayuxs5xcxhylqcf527bit5qn72iifrkrtdg5vcb4qxaihhkiice
+valory/skills/oracle_deployment_abci,bafybeie45aclzfigqr5f32qp2gvh42se44obakser4cq3qz3sh52qa4jpe
+valory/skills/price_estimation_abci,bafybeigyjtc2n7bm7s2s4gsbqh4ftwczmtivgt5zp6n7mhyks5rbsah2hq
+valory/skills/register_reset_abci,bafybeidy6tn6gqkz2jcjspya2sy4tudwzuhmvem7hghtdfoay4rvgll6ua
+valory/skills/registration_abci,bafybeib54vcjymnqejppq6xnjvwvrcrs6ugq64prsgjo65qs6qvdn6u42a
+valory/skills/reset_pause_abci,bafybeigntwmltw62wtyjn4a7vumv7dreyuxng4kfv4lw76ktqswy3rjgum
+valory/skills/safe_deployment_abci,bafybeidjara6duo3cylvq4doewhqsvsjbpesva42nkxrlyqjolkbiaomte
+valory/skills/simple_abci,bafybeia57aqkoaccn5aoasads63o5rwrxcwttozzkxxyixnczc46fjabde
+valory/skills/test_abci,bafybeigvk7tvoiwt2tisirk3ubb22k4kevjv3tngd7pazzzkdrxjzlr6lu
+valory/skills/transaction_settlement_abci,bafybeic5fit66hxn62vd3egbh66xpmvpsz5bu53kvuyq4iyq4brfyge7ri
