+++ packages/valory/agents/hello_world/aea-config.yaml
@@ -18,8 +18,8 @@ protocols:
 - valory/http:1.0.0:bafybeienmids6f5pcey2bnywmiup2faawqgmdhhz65qfciuhkntugztvpi
 skills:
 - valory/abstract_abci:0.1.0:bafybeih62tn62c33bcffjrlxvvokvorli2ow4v6myij352bl3im5dhvloa
-- valory/abstract_round_abci:0.1.0:bafybeievrok3xa3foeay3d3xoscunhbgnqiu6jyse6yck4td76dxsbvsiu
-- valory/hello_world_abci:0.1.0:bafybeifp2vpwiapnusrrpi6mjedxgmovodyrdc7mgvqezhq3xcaapaqfom
+- valory/abstract_round_abci:0.1.0:bafybeihcarapqbgd7czlfyygkdjvb3zxhfkxvud7hd5owajkcwqsxa63gy
+- valory/hello_world_abci:0.1.0:bafybeiflibb7eiydyxuoyvsasc4sq5n3cpyplboprkngu5yhqsncfaicky
 default_ledger: ethereum
 required_ledgers:
 - ethereum
