+++ packages/valory/services/apy_estimation_hardhat/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeibqyfur7wusrbfga66hcbauxvhs2g3j5yhknsxlvyh6mil57aajyu
 fingerprint_ignore_patterns: []
-agent: valory/apy_estimation:0.1.0:bafybeiacs2kuowwshc7zftf6u5uy5g5iftqp72br75qgd3ey7hczjvftm4
+agent: valory/apy_estimation:0.1.0:bafybeiemzfzmgo5sk222miw4zwc2gnt7wdqsub6ddguwrfukq3pwiaubci
 network: hardhat
 number_of_agents: 4
 ---
