+++ packages/valory/services/oracle_goerli/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeig4wu7inb6ibwfgged3djyc7emopcg5n7d7gwrlotne53pp3jfgq4
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeibjnqewgxhxzbqafvqnxmq2ilwhfv6brhz5wtjypmjg74heuyfpgq
+agent: valory/oracle:0.1.0:bafybeiggp7g3reb3q3hisvgjqwgvforbhz26caraqi5wi4cqzny5uiezcy
 network: goerli
 number_of_agents: 4
 ---
