+++ packages/valory/services/oracle_ropsten/service.yaml
@@ -7,7 +7,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeifpb54id7wln2ayku65s3qmih3aljf23rfpsggjwuzc7cnkfroi7y
 fingerprint_ignore_patterns: []
-agent: valory/oracle:0.1.0:bafybeibjnqewgxhxzbqafvqnxmq2ilwhfv6brhz5wtjypmjg74heuyfpgq
+agent: valory/oracle:0.1.0:bafybeiggp7g3reb3q3hisvgjqwgvforbhz26caraqi5wi4cqzny5uiezcy
 network: ropsten
 number_of_agents: 4
 ---
