+++ packages/valory/services/register_reset/service.yaml
@@ -1,7 +1,7 @@
 name: register_reset
 author: valory
 version: 0.1.0
-agent: valory/register_reset:0.1.0:bafybeiedssyqltgi3wvvrvac7s3zmwtnmpezk5qlxtraw2konafzf6bktq
+agent: valory/register_reset:0.1.0:bafybeifkvg2paaso5bqtwsrs4623g55j6wtslop74rjhyvrw2vteatbl7e
 network: hardhat
 number_of_agents: 4
 description: Test and debug tendermint reset mechanism.
